Fix brackets error and move (GM) removal
Fixed the issue where upon saving and reloading the script the ASCII replacements for "(" and ")" would be converted back to the bracket forms. Credit goes to The Aaron, Arcane Scriptomancer for the conversion function.

Also moved the "(GM)" removal from the end of each call to the chat function to be done in the chat function so that it is always done, alongside the splitting of the name to only take the first part.