Add a Utilitites menu and a show parameter
Added the utility functions (rejoin, refresh and help) to the menu command. 
Refresh will only show for the GM.

Added a --show parameter to the menu command, it accepts All, Public, Private, Utilities and Utils.
This is used to filter the menu view to only show certain sets of api buttons.

Added more comments to the script, specifically to the menu function.